+++ src/pages/NewGamePage.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import Header from '../layout/Header';
-import Login from '../Login';
 import Cookies from 'universal-cookie';
 import 'bulma/css/bulma.min.css';
 import {Box} from 'react-bulma-components';
@@ -8,11 +7,15 @@ import NewGame from '../NewGame';
 
 const NewGamePage = () => {
   const cookies = new Cookies();
+  const usernameExists = cookies.get('username') != null;
+
+  if (!usernameExists) {
+    window.location = '/login'
+  }
 
   return (
     <>
       <Header />
-      <Login usernameExists={cookies.get('username') != null} />
       <Box>
         <NewGame />
       </Box>
